petitboot: create petitboot's /var directories at image-build time
On some machines, we see the UI processes start before the discover
server's init script has run. With the petituser functionality, the UI
processes may not be able to create proper log directories if this
happens.

This change creates the /var/log/petitboot and /var/petitboot
directories as part of the image build, rather than on system boot.

  $ cpio -vt < output/images/rootfs.cpio | grep -E 'var/(log/|)petitboot'
  drwxrwxr-x   2 root     2222            0 Jul 19 08:03 var/log/petitboot
  drwxrwxr-x   2 root     2222            0 Jul 19 08:03 var/petitboot

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>