petitboot: Set PATH for plugins from the shell profile
Currently, we expect pb-console to add the pb-plugin binary dir into
PATH.

However, with the petituser separation work, we may escalate to root
with `sudo -i`. This starts a new login shell, so we lose our changes to
PATH.

This change does the PATH additions in the standard shell profile setup,
so that both the pb-console and sudo environments are configured for the
plugin dir.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>